tests: split collector tests per collector type
With getting support for new collectors test_collectors.py was growing
in size, that made it harder to read and maintain.
