+++ tests/unit/test_deps/test_collectors/test_metadata.py
@@ -1,0 +1,50 @@
+from copy import deepcopy
+import json
+
+from pyproject_installer.deps_cmd import deps_command
+
+
+def test_metadata_collector_metadata(deps_data, pyproject_metadata, depsconfig):
+    """Collection of core metadata via prepare_metadata_for_build_wheel"""
+    # prepare source config
+    srcname = "foo"
+    collector = "metadata"
+    input_conf = {"sources": {srcname: {"srctype": collector}}}
+    depsconfig_path = depsconfig(json.dumps(input_conf))
+
+    in_reqs, out_reqs = deps_data
+
+    # configure pyproject with build backend
+    pyproject_metadata(reqs=in_reqs)
+
+    deps_command("sync", depsconfig_path, srcnames=[])
+
+    expected_conf = deepcopy(input_conf)
+    if out_reqs:
+        expected_conf["sources"][srcname]["deps"] = out_reqs
+    actual_conf = json.loads(depsconfig_path.read_text(encoding="utf-8"))
+    assert actual_conf == expected_conf
+
+
+def test_metadata_collector_wheel(
+    deps_data, pyproject_metadata_wheel, depsconfig
+):
+    """Collection of core metadata via build_wheel"""
+    # prepare source config
+    srcname = "foo"
+    collector = "metadata"
+    input_conf = {"sources": {srcname: {"srctype": collector}}}
+    depsconfig_path = depsconfig(json.dumps(input_conf))
+
+    in_reqs, out_reqs = deps_data
+
+    # configure pyproject with build backend
+    pyproject_metadata_wheel(reqs=in_reqs)
+
+    deps_command("sync", depsconfig_path, srcnames=[])
+
+    expected_conf = deepcopy(input_conf)
+    if out_reqs:
+        expected_conf["sources"][srcname]["deps"] = out_reqs
+    actual_conf = json.loads(depsconfig_path.read_text(encoding="utf-8"))
+    assert actual_conf == expected_conf
